Polling for F24 attachments
