+++ src/features/messages/types/liveModeMessages.ts
@@ -1,0 +1,8 @@
+import * as t from "io-ts";
+
+export const LiveModeMessages = t.interface({
+  // interval between updates in millis
+  interval: t.number,
+  // number of new messages
+  count: t.number
+});
