+++ src/types/allowRandomValue.ts
@@ -1,0 +1,3 @@
+import * as t from "io-ts";
+
+export const AllowRandomValue = t.interface({ allowRandomValues: t.boolean });
