+++ src/types/httpResponseCode.ts
@@ -1,0 +1,11 @@
+import * as t from "io-ts";
+
+/* general http response codes */
+export const HttpResponseCode = t.union([
+  t.literal(200),
+  t.literal(400),
+  t.literal(401),
+  t.literal(404),
+  t.literal(429),
+  t.literal(500)
+]);
