+++ src/utils/payment.ts
@@ -3,6 +3,8 @@ import { PaymentDataWithRequiredPayee } from "../../generated/definitions/backen
 import { PaymentProblemJson } from "../../generated/definitions/backend/PaymentProblemJson";
 import { RptId } from "../../generated/definitions/backend/RptId";
 import { NotificationPaymentInfo } from "../features/pn/types/notificationPaymentInfo";
+import { ServicePublic } from "../../generated/definitions/backend/ServicePublic";
+import { PaymentData } from "../../generated/definitions/backend/PaymentData";
 
 export const enum CreditCardBrandEnum {
   "VISAELECTRON" = "VISAELECTRON",
@@ -65,6 +67,11 @@ export const rptIdFromNotificationPaymentInfo = (
   notificationPaymentInfo: NotificationPaymentInfo
 ): RptId =>
   `${notificationPaymentInfo.creditorTaxId}${notificationPaymentInfo.noticeCode}`;
+
+export const rptIdFromServiceAndPaymentData = (
+  service: ServicePublic,
+  paymentData: PaymentData
+) => `${service.organization_fiscal_code}${paymentData.notice_number}`;
 
 export const isPaid = (paymentProblemJSON: PaymentProblemJson) =>
   pipe(
